Extract code from DynavityApp into MainView
